pyproject.toml
[project]
name = "topic-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.88.0",
    "bertopic>=0.17.4",
    "hdbscan>=0.8.42",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "umap-learn>=0.5.11",
]
